Adding PopUp edit

+++ src/components/ListItem.tsx
@@ -24,7 +24,32 @@ const ListItem: React.FC<ListItemProps> = (props: ListItemProps) => {
 
   //Llama al edit<item> de la página padre
   function editItem(){
-    props.handleEdit(props.id);
+    presentAlert({
+      header: "¿Desea modificar: "+ props.text +"?",
+      inputs: [
+        {
+          placeholder: 'Color',
+        },
+        {
+          type: 'number',
+          placeholder: 'Cantidad',
+          min: 0,
+          max: 50,
+        },
+      ],
+      buttons: [
+
+        {
+          text: "Cancelar",
+          cssClass: "alert-button-cancel",
+        },
+        {
+          text: "Modificar",
+          cssClass: "alert-button-confirm",
+          handler: () => {props.handleEdit(props.id)}
+        }
+      ],
+    })
   }
   //Llama al delete<item> de la página padre
   function deleteItem(){
@@ -53,21 +78,44 @@ const ListItem: React.FC<ListItemProps> = (props: ListItemProps) => {
     return <></>
   }
 
-  return (     
-    <IonItem key={props.id} class="remove-padding custom-padding" >
-      <IonItem lines="none" class="remove-padding full-width" href={href}>
-        {quantity()}
-        <IonImg class="pictogram-on-button" src={pictogram} />
-        <IonLabel class='ion-text-wrap'> {props.text}</IonLabel>
-      </IonItem>
-      <IonItem lines='none' slot='end' class='remove-padding fit-width'>
+  const editButton = () => {
+    if(props.handleEdit != undefined) {
+      return (
         <IonButton class='icon-button' icon-only item-end fill='clear' onClick={editItem}>
           <IonIcon icon={createOutline}></IonIcon>
-        </IonButton>    
+        </IonButton>
+      )
+    }
+    return <></>
+  }
+
+  const deleteButton = () => {
+    if(props.handleDelete != undefined) {
+      return (
         <IonButton class='icon-button' icon-only item-end fill='clear' onClick={deleteItem}>
           <IonIcon icon={trashOutline}></IonIcon>
         </IonButton>
+      )
+    }
+  }
+
+  const handleButtons = () => {
+    return (
+      <IonItem lines='none' slot='end' class='remove-padding fit-width'>
+        {editButton()}
+        {deleteButton()}
       </IonItem>
+    )
+  }
+
+  return (     
+    <IonItem key={props.id} class="remove-padding custom-padding" >
+      <IonItem lines="none" class="remove-padding full-width" href={href}>
+        <IonImg class="pictogram-on-button" src={pictogram} />
+        <IonLabel class='ion-text-wrap'> {props.text}</IonLabel>
+        <IonLabel class='quantity'>{quantity()}</IonLabel>
+      </IonItem>
+      {handleButtons()}
     </IonItem>
   )
 }
